Updated CPU code with v2
Added new version of CPU subgrid calculator. This version both speeds up the old code by around 4x and reduces memory usage by half (at least in the example). I have also added in the lookup table reducer so that does not have to be read in on a separate command.